refactor to add probe and snapshot as step 2
